move seeder to another package to avoid import cycles
